requirements.txt
# RedForge CLI Dependencies
# Reference: cid-roadmap-v1 Sprint S-1
# Updated to match pyproject.toml versions

# Core dependencies
click>=8.1.7
pydantic>=2.5.0
httpx>=0.26.0
rich>=13.7.0
typer>=0.9.0
openai>=1.6.0
tqdm>=4.66.0
jinja2>=3.1.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3
reportlab>=4.0.0

# Development dependencies (optional)
pytest>=7.4.3;extra=="dev"
pytest-cov>=4.1.0;extra=="dev"
pytest-asyncio>=0.23.2;extra=="dev"
black>=23.12.0;extra=="dev"
isort>=5.13.0;extra=="dev"
flake8>=6.1.0;extra=="dev"
mypy>=1.8.0;extra=="dev"